feat: implement disconnect method for WS
The commit title is a bit misleading. A huge part of this commit was
spent on creating a channel between WS and Redis worker threads. This
channel allowed to not send WS session as mutex to the redis worker. As
a result the implementaiton is a bit more straighforward (at least in my
opinion).

The one thing that needs tweaking is figuring out how to deal with
disconnect regarding the channel itself. Right now I just ignore it,
which is not the best solution. :)